+++ stories/02-Design System/04-DataEntry/Select/examples/download.tsx
@@ -1,7 +1,7 @@
-import React, {useState} from 'react';
+import {useState} from 'react';
 import {KitSelect} from '@kit/DataEntry/';
-import {KitDivider, KitSpace} from '@kit/Layout/';
-import {KitButton} from '@kit/General/';
+import {KitSpace} from '@kit/Layout/';
+import {KitButton, KitTypography} from '@kit/General/';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faFile} from '@fortawesome/free-regular-svg-icons';
 import {faDownload} from '@fortawesome/free-solid-svg-icons';
@@ -52,43 +52,47 @@ const App = () => {
     const style = {width: 320};
 
     return (
-        <KitSpace direction="vertical" size={2}>
-            <KitDivider orientation="left">Download compact</KitDivider>
-            <KitSpace.Compact>
-                <KitSelect
-                    placeholder="Select a file to download"
-                    style={style}
-                    onChange={handleChange}
-                    options={iconOptions}
-                />
-                {selectedOption && (
-                    <KitButton
-                        type="primary"
-                        icon={<FontAwesomeIcon icon={faDownload} />}
-                        iconSize="xl"
-                        onClick={handleClick}
-                    />
-                )}
-            </KitSpace.Compact>
-            <KitDivider orientation="left">Download under the select</KitDivider>
+        <KitSpace direction="vertical" size={'l'}>
             <KitSpace direction="vertical">
-                <KitSelect
-                    placeholder="Select a file to download"
-                    style={style}
-                    onChange={handleChange2}
-                    options={iconOptions}
-                />
-                {selectedOption2 && (
-                    <KitButton
-                        type="primary"
-                        icon={<FontAwesomeIcon icon={faDownload} />}
-                        onClick={handleClick2}
-                        block
-                        style={{marginTop: '16px'}}
-                    >
-                        Download
-                    </KitButton>
-                )}
+                <KitTypography.Text weight="bold">Download compact</KitTypography.Text>
+                <KitSpace.Compact>
+                    <KitSelect
+                        placeholder="Select a file to download"
+                        style={style}
+                        onChange={handleChange}
+                        options={iconOptions}
+                    />
+                    {selectedOption && (
+                        <KitButton
+                            type="primary"
+                            icon={<FontAwesomeIcon icon={faDownload} />}
+                            iconSize="xl"
+                            onClick={handleClick}
+                        />
+                    )}
+                </KitSpace.Compact>
+            </KitSpace>
+            <KitSpace direction="vertical">
+                <KitTypography.Text weight="bold">Download under the select</KitTypography.Text>
+                <KitSpace direction="vertical">
+                    <KitSelect
+                        placeholder="Select a file to download"
+                        style={style}
+                        onChange={handleChange2}
+                        options={iconOptions}
+                    />
+                    {selectedOption2 && (
+                        <KitButton
+                            type="primary"
+                            icon={<FontAwesomeIcon icon={faDownload} />}
+                            onClick={handleClick2}
+                            block
+                            style={{marginTop: '16px'}}
+                        >
+                            Download
+                        </KitButton>
+                    )}
+                </KitSpace>
             </KitSpace>
         </KitSpace>
     );