+++ stories/02-Design System/04-DataEntry/Select/examples/index.tsx
@@ -1,4 +1,5 @@
 import Basic from './basic';
+import CustomRender from './customRender';
 import Disabled from './disabled';
 import Download from './download';
 import Group from './group';
@@ -7,6 +8,7 @@ import Wrapper from './wrapper';
 import Multipleandtags from './multipleandtags';
 
 import BasicSource from './basic?raw';
+import CustomRenderSource from './customRender?raw';
 import DisabledSource from './disabled?raw';
 import DownloadSource from './download?raw';
 import GroupSource from './group?raw';
@@ -16,6 +18,7 @@ import MultipleandtagsSource from './multipleandtags?raw';
 
 export const Sources = {
     Basic: BasicSource,
+    CustomRender: CustomRenderSource,
     Disabled: DisabledSource,
     Download: DownloadSource,
     Group: GroupSource,
@@ -25,6 +28,7 @@ export const Sources = {
 };
 export default {
     Basic,
+    CustomRender,
     Disabled,
     Download,
     Group,