+++ stories/02-Design System/04-DataEntry/Select/examples/multipleandtags.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
 import {KitSelect} from '@kit/DataEntry/';
-import {KitDivider, KitSpace} from '@kit/Layout/';
+import {KitSpace} from '@kit/Layout/';
+import {KitTypography} from '@kit/General';
 import {useKitTheme} from '@theme/useKitTheme';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faFile} from '@fortawesome/free-regular-svg-icons';
@@ -61,47 +61,51 @@ const App = () => {
     const style = {width: 340};
 
     return (
-        <KitSpace direction="vertical">
-            <KitDivider orientation="left">Mode=multiple (default / oneLine)</KitDivider>
-            <KitSpace>
-                <KitSelect
-                    mode="multiple"
-                    placeholder="Select a status"
-                    style={style}
-                    defaultActiveFirstOption={false}
-                    onChange={handleChange}
-                    options={colorOptions}
-                />
-                <KitSelect
-                    mode="multiple"
-                    placeholder="Select a status"
-                    style={style}
-                    oneLineTags
-                    defaultActiveFirstOption={false}
-                    onChange={handleChange}
-                    options={colorOptions}
-                />
+        <KitSpace direction="vertical" size={'l'}>
+            <KitSpace direction="vertical">
+                <KitTypography.Text weight="bold">Mode=multiple (default / oneLine)</KitTypography.Text>
+                <KitSpace>
+                    <KitSelect
+                        mode="multiple"
+                        placeholder="Select a status"
+                        style={style}
+                        defaultActiveFirstOption={false}
+                        onChange={handleChange}
+                        options={colorOptions}
+                    />
+                    <KitSelect
+                        mode="multiple"
+                        placeholder="Select a status"
+                        style={style}
+                        oneLineTags
+                        defaultActiveFirstOption={false}
+                        onChange={handleChange}
+                        options={colorOptions}
+                    />
+                </KitSpace>
             </KitSpace>
-            <KitDivider orientation="left">Mode=tags (default / oneLine)</KitDivider>
-            <KitSpace>
-                <KitSelect
-                    mode="tags"
-                    placeholder="Select a file"
-                    style={style}
-                    defaultActiveFirstOption={false}
-                    onChange={handleChange}
-                    options={iconOptions}
-                />
-                <KitSelect
-                    mode="tags"
-                    placeholder="Select a file"
-                    style={style}
-                    labelOnly
-                    oneLineTags
-                    defaultActiveFirstOption={false}
-                    onChange={handleChange}
-                    options={iconOptions}
-                />
+            <KitSpace direction="vertical">
+                <KitTypography.Text weight="bold">Mode=tags (default / oneLine)</KitTypography.Text>
+                <KitSpace>
+                    <KitSelect
+                        mode="tags"
+                        placeholder="Select a file"
+                        style={style}
+                        defaultActiveFirstOption={false}
+                        onChange={handleChange}
+                        options={iconOptions}
+                    />
+                    <KitSelect
+                        mode="tags"
+                        placeholder="Select a file"
+                        style={style}
+                        labelOnly
+                        oneLineTags
+                        defaultActiveFirstOption={false}
+                        onChange={handleChange}
+                        options={iconOptions}
+                    />
+                </KitSpace>
             </KitSpace>
         </KitSpace>
     );