+++ stories/02-Design System/05-DataDisplay/Avatar/examples/index.tsx
@@ -2,6 +2,7 @@ import Badge from './badge';
 import Basic from './basic';
 import Color from './color';
 import Initials from './initials';
+import ImageFit from './imageFit';
 import Group from './group';
 import Type from './type';
 import Responsive from './responsive';
@@ -10,6 +11,7 @@ import BadgeSource from './badge?raw';
 import BasicSource from './basic?raw';
 import ColorSource from './color?raw';
 import InitialsSource from './initials?raw';
+import ImageFitSource from './imageFit?raw';
 import GroupSource from './group?raw';
 import TypeSource from './type?raw';
 import ResponsiveSource from './responsive?raw';
@@ -19,6 +21,7 @@ export const Sources = {
     Basic: BasicSource,
     Color: ColorSource,
     Initials: InitialsSource,
+    ImageFit: ImageFitSource,
     Group: GroupSource,
     Type: TypeSource,
     Responsive: ResponsiveSource
@@ -28,6 +31,7 @@ export default {
     Basic,
     Color,
     Initials,
+    ImageFit,
     Group,
     Type,
     Responsive